hydrate vault view (#1050)
* hydrate vault view

* fix error

* fix error with duplicated keys

* style fix

* fix icon
