+++ pages/owner/[address]/index.tsx
@@ -3,7 +3,7 @@ import { WithConnection } from 'components/connectWallet/ConnectWallet'
 import { AppLayout } from 'components/Layouts'
 import { getAddress } from 'ethers/lib/utils'
 import { WithTermsOfService } from 'features/termsOfService/TermsOfService'
-import { VaultsOverviewView } from 'features/vaultsOverview/VaultsOverviewView'
+import { VaultsOverviewView } from 'features/vaultsOverview/VaultOverviewView'
 import { WithLoadingIndicator } from 'helpers/AppSpinner'
 import { WithErrorHandler } from 'helpers/errorHandlers/WithErrorHandler'
 import { useObservable } from 'helpers/observableHook'
